gmail-automation-agent/requirements.txt
Flask>=2.0
google-auth>=2.0
google-auth-oauthlib>=0.7
google-api-python-client>=2.0
openai>=1.0
python-dotenv>=0.21